refactor: rename waltti config file
