+++ app/configurations/config.jyvaskyla.js
@@ -5,7 +5,7 @@ const CONFIG = 'jyvaskyla';
 const APP_TITLE = 'Reittiopas Jyväskylä';
 const APP_DESCRIPTION = 'Jyväskylän uusi reittiopas';
 
-const walttiConfig = require('./waltti').default;
+const walttiConfig = require('./config.waltti').default;
 
 const minLat = 61.835318;
 const maxLat = 62.603473;
