+++ app/configurations/config.mikkeli.js
@@ -5,7 +5,7 @@ const CONFIG = 'mikkeli';
 const APP_TITLE = 'Mikkelin Reittiopas';
 const APP_DESCRIPTION = 'Mikkelin julkisen liikenteen reittiopas';
 
-const walttiConfig = require('./waltti').default;
+const walttiConfig = require('./config.waltti').default;
 
 const minLat = 61.65922;
 const maxLat = 61.717372;
