+++ app/configurations/config.rovaniemi.js
@@ -5,7 +5,7 @@ const CONFIG = 'rovaniemi';
 const APP_TITLE = 'Rovaniemen reittiopas';
 const APP_DESCRIPTION = 'Rovaniemen uusi reittiopas';
 
-const walttiConfig = require('./waltti').default;
+const walttiConfig = require('./config.waltti').default;
 
 const minLat = 66.147037;
 const maxLat = 67.180128;
